web_car/requirements.txt
pyknp==0.4.1
bert-serving-client==1.10.0
elasticsearch==7.0.4
Flask==1.1.1
